requirements.txt
fastapi 
uvicorn
lxml
semver
dotenv
python-gitlab
requests
pyyaml
GitPython
aiohttp